Moved around networkConfig and added unit tests
