+++ .travis.yml
@@ -14,4 +14,4 @@ script:
   - fgt go vet ./...
   - fgt golint ./...
   - go test -bench . -benchmem ./...
-  - bin/coverage
+  - bin/coverage travis
